Move BlockSlicer somewhere else
