+++ include/openPMD/auxiliary/BlockSlicer.hpp
@@ -1,4 +1,4 @@
-/* Copyright 2025 Franz Poeschel
+/* Copyright 2018-2021 Franz Poeschel
  *
  * This file is part of openPMD-api.
  *
@@ -19,13 +19,35 @@
  * If not, see <http://www.gnu.org/licenses/>.
  */
 
-/* Legacy header for backward compatibility */
-
 #pragma once
 
-#include "openPMD/auxiliary/BlockSlicer.hpp"
+#include "openPMD/Dataset.hpp"
 
-namespace openPMD
+#include <memory>
+
+namespace openPMD::auxiliary
 {
-using auxiliary::BlockSlicer;
-}
+/**
+ * Abstract class to associate a thread with its local cuboid in the total
+ * cuboid.
+ */
+class BlockSlicer
+{
+public:
+    /**
+     * Associate the current thread with its cuboid.
+     * @param totalExtent The total extent of the cuboid.
+     * @param size The number of threads to be used (not greater than MPI size).
+     * @param rank The MPI rank.
+     * @return A pair of the cuboid's offset and extent.
+     */
+    virtual std::pair<Offset, Extent>
+    sliceBlock(Extent &totalExtent, int size, int rank) = 0;
+
+    virtual std::unique_ptr<BlockSlicer> clone() const = 0;
+
+    /** This class will be derived from
+     */
+    virtual ~BlockSlicer() = default;
+};
+} // namespace openPMD::auxiliary
